Add multicast message caching function.
It is difficult to deal with multicast messages which represent some state
information. When a process suscribes a multicast message group, there will be
no way to verify the state until receiving the next message of that group which
maybe never come.

To slove this problem, ipcon driver caches the last multicast message for every
group. When some process suscribes a multicast group (call ipcon_join_group()),
it will send the cached message back and ipcon_join_group() queues it for next
reading. In this way, a multicast message group suscriber will "always" get the
newest state information.